@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute, Params, Router} from '@angular/router';
-import {FormArray, FormControl, FormGroup, Validators} from '@angular/forms';
+import {UntypedFormArray, UntypedFormControl, UntypedFormGroup, Validators} from '@angular/forms';
 
 import {RecipeService} from '../recipe.service';
 
@@ -13,7 +13,7 @@ export class RecipeEditComponent implements OnInit {
 
   id: number;
   editMode = false;
-  recipeForm: FormGroup;
+  recipeForm: UntypedFormGroup;
 
   constructor(
     private route: ActivatedRoute,
@@ -37,7 +37,7 @@ export class RecipeEditComponent implements OnInit {
     let recipeName = '';
     let recipeImagePath = '';
     let recipeDescription = '';
-    const recipeIngredients = new FormArray([]);
+    const recipeIngredients = new UntypedFormArray([]);
 
     if (this.editMode) {
       const recipe = this.recipeService.getRecipeByID(this.id);
@@ -48,10 +48,10 @@ export class RecipeEditComponent implements OnInit {
       if (recipe.ingredients) {
         for (const ingredient of recipe.ingredients) {
           recipeIngredients.push(
-            new FormGroup({
+            new UntypedFormGroup({
 
-              name: new FormControl(ingredient.name, Validators.required),
-              amount: new FormControl(ingredient.amount, [
+              name: new UntypedFormControl(ingredient.name, Validators.required),
+              amount: new UntypedFormControl(ingredient.amount, [
                 Validators.required,
                 Validators.pattern(/^[1-9]+[0-9]*$/)
               ])
@@ -61,19 +61,19 @@ export class RecipeEditComponent implements OnInit {
       }
     }
 
-    this.recipeForm = new FormGroup({
-      name: new FormControl(recipeName, Validators.required),
-      imagePath: new FormControl(recipeImagePath, Validators.required),
-      description: new FormControl(recipeDescription, Validators.required),
+    this.recipeForm = new UntypedFormGroup({
+      name: new UntypedFormControl(recipeName, Validators.required),
+      imagePath: new UntypedFormControl(recipeImagePath, Validators.required),
+      description: new UntypedFormControl(recipeDescription, Validators.required),
       ingredients: recipeIngredients
     });
   }
 
   onAddIngredient() {
-    (this.recipeForm.get('ingredients') as FormArray).push(
-      new FormGroup({
-        name: new FormControl(null, Validators.required),
-        amount: new FormControl(null, [
+    (this.recipeForm.get('ingredients') as UntypedFormArray).push(
+      new UntypedFormGroup({
+        name: new UntypedFormControl(null, Validators.required),
+        amount: new UntypedFormControl(null, [
           Validators.required,
           Validators.pattern(/^[1-9]+[0-9]*$/)
         ])
@@ -82,7 +82,7 @@ export class RecipeEditComponent implements OnInit {
   }
 
   get controls() { // a getter!
-    return (this.recipeForm.get('ingredients') as FormArray).controls;
+    return (this.recipeForm.get('ingredients') as UntypedFormArray).controls;
   }
 
   onSubmit() {
@@ -117,6 +117,6 @@ export class RecipeEditComponent implements OnInit {
   }
 
   onDeleteIngredient(index: number) {
-    (this.recipeForm.get('ingredients') as FormArray).removeAt(index);
+    (this.recipeForm.get('ingredients') as UntypedFormArray).removeAt(index);
   }
 }
